C-WCOW: Securitpolicy pkg changes and enforce CIMs
