+++ internal/protocol/guestresource/resources.go
@@ -115,9 +115,10 @@ type BlockCIMDevice struct {
 
 type WCOWBlockCIMMounts struct {
 	// BlockCIMs should be ordered from merged CIM followed by Layer n .. layer 1
-	BlockCIMs  []BlockCIMDevice `json:"BlockCIMs,omitempty"`
-	VolumeGuid guid.GUID        `json:"VolumeGuid,omitempty"`
-	MountFlags uint32           `json:"MountFlags,omitempty"`
+	BlockCIMs   []BlockCIMDevice `json:"BlockCIMs,omitempty"`
+	VolumeGuid  guid.GUID        `json:"VolumeGuid,omitempty"`
+	MountFlags  uint32           `json:"MountFlags,omitempty"`
+	ContainerID string
 }
 
 type WCOWMappedVirtualDisk struct {
