+++ pkg/securitypolicy/open_door.rego
@@ -5,6 +5,7 @@ api_version := "@@API_VERSION@@"
 mount_device := {"allowed": true}
 mount_overlay := {"allowed": true}
 create_container := {"allowed": true, "env_list": null, "allow_stdio_access": true}
+mount_cims := {"allowed": true}
 unmount_device := {"allowed": true}
 unmount_overlay := {"allowed": true}
 exec_in_container := {"allowed": true, "env_list": null}
