+++ pkg/securitypolicy/policy.rego
@@ -9,6 +9,7 @@ mount_device := data.framework.mount_device
 unmount_device := data.framework.unmount_device
 mount_overlay := data.framework.mount_overlay
 unmount_overlay := data.framework.unmount_overlay
+mount_cims:= data.framework.mount_cims
 create_container := data.framework.create_container
 exec_in_container := data.framework.exec_in_container
 exec_external := data.framework.exec_external
