--- pkg/securitypolicy/securitypolicy_uvmpath_linux.go
@@ -1,34 +1,0 @@
-//go:build linux
-// +build linux
-
-package securitypolicy
-
-import (
-	"strings"
-
-	specInternal "github.com/Microsoft/hcsshim/internal/guest/spec"
-	"github.com/Microsoft/hcsshim/internal/guestpath"
-)
-
-// This is being used by StandEnforcer.
-// substituteUVMPath substitutes mount prefix to an appropriate path inside
-// UVM. At policy generation time, it's impossible to tell what the sandboxID
-// will be, so the prefix substitution needs to happen during runtime.
-func substituteUVMPath(sandboxID string, m mountInternal) mountInternal {
-	if strings.HasPrefix(m.Source, guestpath.SandboxMountPrefix) {
-		m.Source = specInternal.SandboxMountSource(sandboxID, m.Source)
-	} else if strings.HasPrefix(m.Source, guestpath.HugePagesMountPrefix) {
-		m.Source = specInternal.HugePagesMountSource(sandboxID, m.Source)
-	}
-	return m
-}
-
-// SandboxMountsDir returns sandbox mounts directory inside UVM/host.
-func SandboxMountsDir(sandboxID string) string {
-	return specInternal.SandboxMountsDir((sandboxID))
-}
-
-// HugePagesMountsDir returns hugepages mounts directory inside UVM.
-func HugePagesMountsDir(sandboxID string) string {
-	return specInternal.HugePagesMountsDir(sandboxID)
-}
