+++ pkg/securitypolicy/version_api
@@ -1,1 +1,1 @@
-0.10.0
+0.11.0
