+++ pkg/securitypolicy/version_framework
@@ -1,1 +1,1 @@
-0.3.0
+0.4.0
